Rename `python_api.py` to `approx.py`
The `python_api.py` file used to also contain the implementation of
`raises`, but that was moved to its own file a while ago. Now
`python_api.py` only houses the `approx` code, so let's rename it to the
more obvious and less generic `approx.py`.
